Extract step handler
