+++ controllers/apply/form-router-next/views/terms.njk
@@ -19,7 +19,7 @@
 
             {{ formErrors( __('global.misc.errorTitle'), errors) }}
 
-            <form method="post" action="{{ formBaseUrl }}/submit" class="u-constrained-content-wide">
+            <form method="post" action="{{ formBaseUrl }}/submission" class="u-constrained-content-wide">
 
                 <div class="s-prose u-text-medium u-margin-bottom-l">
                     <p>By submitting an application to The National Lottery Community Fund, the organisation named in the application (referred to as “you” in these Terms and Conditions) agrees, if awarded a grant, to:</p>
